requirements.txt
# Core dependencies
tree_sitter
tree_sitter_languages
langchain
langchain-openai
langchain-community
openai
ollama
sentence-transformers
torch
transformers
streamlit
watchdog
gitpython
chromadb
astor
faiss-cpu
prompt_toolkit
simple-term-menu
